feat: interfaces-b12-cli-workflow-mcp-006 - Preserve MCP serve, protocol, and resume behavior
