+++ pkg/transports/cli/climanifestparity/execution_parity_test.go
@@ -8,6 +8,8 @@ import (
 	"io"
 	"net/http"
 	"net/http/httptest"
+	"os"
+	"path/filepath"
 	"runtime"
 	"strings"
 	"testing"
@@ -15,13 +17,237 @@ import (
 	"github.com/portpowered/infinite-you/pkg/root"
 	"github.com/portpowered/infinite-you/pkg/testutil"
 	"github.com/portpowered/infinite-you/pkg/transports/cli"
+	"github.com/portpowered/infinite-you/pkg/transports/cli/cliinputs"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/climanifest"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/climanifestparity"
-	"github.com/portpowered/infinite-you/pkg/transports/cli/cliinputs"
 	modelscli "github.com/portpowered/infinite-you/pkg/transports/cli/models"
 	sessioncli "github.com/portpowered/infinite-you/pkg/transports/cli/session"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
+	"github.com/spf13/cobra"
 )
+
+func TestGeneratedVsLegacyParityMatrix_WorkflowMCPFamily(t *testing.T) {
+	cases := []struct {
+		commandID string
+		path      string
+	}{
+		{commandID: "you.mcp", path: "you mcp"},
+		{commandID: "you.mcp.serve", path: "you mcp serve"},
+		{commandID: "you.workflow.preview", path: "you workflow preview"},
+		{commandID: "you.workflow.validate", path: "you workflow validate"},
+	}
+
+	for _, comparison := range []struct {
+		name string
+		run  func(*testing.T, string, string, *cobra.Command, *cobra.Command)
+	}{
+		{name: "identity", run: func(t *testing.T, commandID, path string, legacy, generated *cobra.Command) {
+			legacyCmd, err := climanifestparity.FindCommandByPath(legacy, path)
+			if err != nil {
+				t.Fatalf("legacy FindCommandByPath(%q) error = %v", path, err)
+			}
+			generatedCmd, err := climanifestparity.FindCommandByPath(generated, path)
+			if err != nil {
+				t.Fatalf("generated FindCommandByPath(%q) error = %v", path, err)
+			}
+			assertNoConstructorMismatches(t, climanifestparity.CompareConstructorIdentityParity(commandID, legacyCmd, generatedCmd))
+		}},
+		{name: "help", run: func(t *testing.T, commandID, path string, legacy, generated *cobra.Command) {
+			mismatches, err := climanifestparity.CompareConstructorHelpParity(commandID, legacy, generated, path)
+			if err != nil {
+				t.Fatalf("CompareConstructorHelpParity(%q) error = %v", path, err)
+			}
+			assertNoConstructorMismatches(t, mismatches)
+		}},
+		{name: "completion", run: func(t *testing.T, commandID, path string, legacy, generated *cobra.Command) {
+			mismatches, err := climanifestparity.CompareConstructorCompletionInventoryParity(commandID, path, legacy, generated)
+			if err != nil {
+				t.Fatalf("CompareConstructorCompletionInventoryParity(%q) error = %v", path, err)
+			}
+			assertNoConstructorMismatches(t, mismatches)
+		}},
+	} {
+		t.Run(comparison.name, func(t *testing.T) {
+			for _, tc := range cases {
+				t.Run(tc.commandID, func(t *testing.T) {
+					legacy, generated := mustWorkflowMCPConstructorRoots(t)
+					comparison.run(t, tc.commandID, tc.path, legacy, generated)
+				})
+			}
+		})
+	}
+
+	parsingCases := []struct {
+		name        string
+		commandID   string
+		argv        []string
+		flagChecks  []string
+		wantErr     bool
+		errContains string
+	}{
+		{name: "mcp defaults", commandID: "you.mcp.serve", argv: []string{"mcp", "serve"}, flagChecks: []string{"runtime", "fixture-catalog", "project-root"}},
+		{name: "mcp explicit runtime", commandID: "you.mcp.serve", argv: []string{"mcp", "serve", "--runtime", "--project-root", "project"}, flagChecks: []string{"runtime", "project-root"}},
+		{name: "workflow validate source", commandID: "you.workflow.validate", argv: []string{"workflow", "validate", "--kind", "WORKFLOW_NAME", "--value", "review"}, flagChecks: []string{"kind", "value", "dir"}},
+		{name: "workflow preview inline", commandID: "you.workflow.preview", argv: []string{"--json", "workflow", "preview", "--kind", "INLINE_WORKFLOW", "--inline", "phase('setup');"}, flagChecks: []string{"json", "kind", "inline"}},
+		{name: "unknown workflow flag", commandID: "you.workflow.validate", argv: []string{"workflow", "validate", "--unknown"}, wantErr: true, errContains: "unknown flag"},
+	}
+	for _, tc := range parsingCases {
+		t.Run("parsing/"+tc.name, func(t *testing.T) {
+			legacy, generated := mustWorkflowMCPConstructorRoots(t)
+			assertNoConstructorMismatches(t, climanifestparity.CompareConstructorParseParity(tc.commandID, legacy, generated, tc.argv, tc.wantErr, tc.errContains))
+			if tc.wantErr {
+				return
+			}
+			legacyLeaf, _, _ := climanifestparity.ParseArgvOnRoot(legacy, tc.argv)
+			generatedLeaf, _, _ := climanifestparity.ParseArgvOnRoot(generated, tc.argv)
+			for _, flag := range tc.flagChecks {
+				assertNoConstructorMismatches(t, climanifestparity.CompareConstructorFlagParity(tc.commandID, flag, legacyLeaf, generatedLeaf))
+			}
+		})
+	}
+}
+
+func mustWorkflowMCPConstructorRoots(t *testing.T) (*cobra.Command, *cobra.Command) {
+	t.Helper()
+	legacy, generated, err := cli.NewWorkflowMCPFamilyParityRoots()
+	if err != nil {
+		t.Fatalf("NewWorkflowMCPFamilyParityRoots() error = %v", err)
+	}
+	return legacy, generated
+}
+
+func TestGeneratedVsLegacyParityMatrix_WorkflowMCPExecution(t *testing.T) {
+	projectRoot := workflowParityProject(t)
+	for _, tc := range workflowExecutionParityCases(projectRoot) {
+		t.Run(tc.name, func(t *testing.T) {
+			legacyRoot, generatedRoot := mustWorkflowMCPConstructorRoots(t)
+			legacyOut, legacyErrOut, legacyErr := executeConstructorRoot(legacyRoot, tc.argv)
+			generatedOut, generatedErrOut, generatedErr := executeConstructorRoot(generatedRoot, tc.argv)
+			assertWorkflowExecutionParity(t, tc, legacyOut, generatedOut, legacyErrOut, generatedErrOut, legacyErr, generatedErr)
+		})
+	}
+}
+
+type workflowExecutionParityCase struct {
+	name            string
+	argv            []string
+	wantErr         bool
+	wantErrContains string
+	wantStdout      []string
+}
+
+func workflowExecutionParityCases(projectRoot string) []workflowExecutionParityCase {
+	validInline := `meta({ name: "review", version: 1 }); phase("setup");`
+	return []workflowExecutionParityCase{
+		{
+			name:       "workflow validate named human success",
+			argv:       []string{"workflow", "validate", "--dir", projectRoot, "--kind", "WORKFLOW_NAME", "--value", " review "},
+			wantStdout: []string{"Workflow validation passed.", ".claude/workflows/review.js", "Source hash:"},
+		},
+		{
+			name:       "workflow validate file JSON success",
+			argv:       []string{"--json", "workflow", "validate", "--dir", projectRoot, "--kind", "WORKFLOW_FILE", "--value", " factory/workflows/review.js "},
+			wantStdout: []string{`"valid":true`, `"sourceRef":"factory/workflows/review.js"`, `"blockingDiagnostics":[]`},
+		},
+		{
+			name:       "workflow preview inline human success",
+			argv:       []string{"workflow", "preview", "--dir", projectRoot, "--kind", "INLINE_WORKFLOW", "--inline", "  " + validInline + "  "},
+			wantStdout: []string{"Factory preview passed.", "Source ref: inline", "Policy hash:", "Result constraints:"},
+		},
+		{
+			name: "workflow preview named JSON success with optional inputs",
+			argv: []string{"--json", "workflow", "preview", "--dir", projectRoot, "--kind", "WORKFLOW_NAME", "--value", "review",
+				"--args-schema", `{"type":"object"}`, "--requested-policy", `{}`},
+			wantStdout: []string{`"valid":true`, `"sourceResolution"`, `"policyPreview"`, `"resultConstraints"`},
+		},
+		{
+			name:            "workflow malformed source failure",
+			argv:            []string{"workflow", "validate", "--dir", projectRoot, "--kind", "INLINE_WORKFLOW", "--inline", "phase("},
+			wantErr:         true,
+			wantErrContains: "workflow validation found blocking issues",
+			wantStdout:      []string{"Workflow validation failed.", "Blocking diagnostics:", "workflow.source.syntaxError"},
+		},
+		{
+			name:            "workflow unresolved source JSON failure",
+			argv:            []string{"--json", "workflow", "validate", "--dir", projectRoot, "--kind", "WORKFLOW_NAME", "--value", "missing"},
+			wantErr:         true,
+			wantErrContains: "workflow validation found blocking issues",
+			wantStdout:      []string{`"valid":false`, `"blockingDiagnostics"`, "workflow.source.notFound"},
+		},
+		{
+			name:            "workflow incomplete source failure",
+			argv:            []string{"workflow", "validate", "--dir", projectRoot, "--kind", "WORKFLOW_NAME"},
+			wantErr:         true,
+			wantErrContains: "value is required when kind is WORKFLOW_NAME",
+		},
+		{
+			name:            "workflow conflicting source failure",
+			argv:            []string{"workflow", "preview", "--dir", projectRoot, "--value", "review", "--inline", validInline},
+			wantErr:         true,
+			wantErrContains: "--inline cannot be used when kind is WORKFLOW_NAME",
+		},
+		{
+			name:            "MCP mutually exclusive source failure",
+			argv:            []string{"mcp", "serve", "--runtime", "--fixture-catalog", "fixtures.json"},
+			wantErr:         true,
+			wantErrContains: "cannot combine --runtime with --fixture-catalog",
+		},
+	}
+}
+
+func workflowParityProject(t *testing.T) string {
+	t.Helper()
+	projectRoot := t.TempDir()
+	for _, relativeDir := range []string{filepath.Join(".claude", "workflows"), filepath.Join("factory", "workflows")} {
+		dir := filepath.Join(projectRoot, relativeDir)
+		if err := os.MkdirAll(dir, 0o755); err != nil {
+			t.Fatalf("create workflow fixture directory: %v", err)
+		}
+		if err := os.WriteFile(filepath.Join(dir, "review.js"), []byte(`meta({ name: "review", version: 1 }); phase("setup");`), 0o600); err != nil {
+			t.Fatalf("write workflow fixture: %v", err)
+		}
+	}
+	return projectRoot
+}
+
+func assertWorkflowExecutionParity(
+	t *testing.T,
+	tc workflowExecutionParityCase,
+	legacyOut, generatedOut, legacyErrOut, generatedErrOut string,
+	legacyErr, generatedErr error,
+) {
+	t.Helper()
+	if (legacyErr != nil) != tc.wantErr || (generatedErr != nil) != tc.wantErr {
+		t.Fatalf("execution errors: legacy=%v generated=%v wantErr=%t", legacyErr, generatedErr, tc.wantErr)
+	}
+	if fmt.Sprint(legacyErr) != fmt.Sprint(generatedErr) {
+		t.Fatalf("exit/error drift: legacy=%v generated=%v", legacyErr, generatedErr)
+	}
+	if tc.wantErrContains != "" && !strings.Contains(fmt.Sprint(legacyErr), tc.wantErrContains) {
+		t.Fatalf("error = %v, want substring %q", legacyErr, tc.wantErrContains)
+	}
+	if legacyOut != generatedOut {
+		t.Fatalf("stdout drift:\nlegacy:\n%s\ngenerated:\n%s", legacyOut, generatedOut)
+	}
+	if legacyErrOut != generatedErrOut {
+		t.Fatalf("stderr drift:\nlegacy:\n%s\ngenerated:\n%s", legacyErrOut, generatedErrOut)
+	}
+	for _, want := range tc.wantStdout {
+		if !strings.Contains(legacyOut, want) {
+			t.Fatalf("stdout missing %q:\n%s", want, legacyOut)
+		}
+	}
+}
+
+func executeConstructorRoot(root *cobra.Command, argv []string) (stdout, stderr string, err error) {
+	var out bytes.Buffer
+	var errOut bytes.Buffer
+	root.SetOut(&out)
+	root.SetErr(&errOut)
+	root.SetArgs(argv)
+	err = root.Execute()
+	return out.String(), errOut.String(), err
+}
 
 func TestProductionManifestCompletionParity_RootAndSessionShow(t *testing.T) {
 	manifestPath := testutil.MustRepoPath(t, climanifest.ProductionManifestPath)
