+++ pkg/transports/cli/root_session_test.go
@@ -2,7 +2,6 @@ package cli
 
 import (
 	"bytes"
-	"context"
 	"io"
 	"os"
 	"path/filepath"
@@ -14,7 +13,6 @@ import (
 	"github.com/portpowered/infinite-you/pkg/factory/sessions/execution/fixtures"
 	workflowsource "github.com/portpowered/infinite-you/pkg/orchestrators/javascript/source"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/session"
-	startupcli "github.com/portpowered/infinite-you/pkg/transports/cli/startup"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/workflow"
 )
 
@@ -705,59 +703,6 @@ func TestWorkflowCommand_LongHelpPresentsCanonicalFactorySessionSuccessorsFirst(
 	}
 	if !strings.Contains(validateCmd.Long, "shared workflow validation contract") {
 		t.Fatalf("validate long help should preserve its validation behavior:\n%s", validateCmd.Long)
-	}
-}
-
-func TestNewWorkflowMCPHandlerRegistriesWiresClassificationSpecificStableIDs(t *testing.T) {
-	registries, err := newWorkflowMCPHandlerRegistries(&cliGlobalOptions{}, RootCommandOptions{})
-	if err != nil {
-		t.Fatalf("newWorkflowMCPHandlerRegistries() error = %v", err)
-	}
-	if _, err := registries.MCP.Lookup("you.mcp.serve"); err != nil {
-		t.Fatalf("MCP.Lookup(you.mcp.serve) error = %v", err)
-	}
-	for _, commandID := range []string{"you.workflow.preview", "you.workflow.validate"} {
-		if _, err := registries.WorkflowCompatibility.Lookup(commandID); err != nil {
-			t.Fatalf("WorkflowCompatibility.Lookup(%q) error = %v", commandID, err)
-		}
-	}
-}
-
-func TestProductionMCPServeGeneratedMetadataDelegatesExistingStartupBoundary(t *testing.T) {
-	var got startupcli.Request
-	startup := func(_ context.Context, request startupcli.Request) error {
-		got = request
-		return nil
-	}
-	stdin := strings.NewReader("protocol input")
-	var stdout bytes.Buffer
-	root := NewRootCommandWithOptions(RootCommandOptions{Startup: startup})
-	root.SetIn(stdin)
-	root.SetOut(&stdout)
-	root.SetErr(io.Discard)
-	root.SetArgs([]string{"mcp", "serve", "--runtime", "--project-root", "project"})
-
-	if err := root.Execute(); err != nil {
-		t.Fatalf("execute generated mcp serve: %v", err)
-	}
-	if got.Kind != startupcli.KindMCPServe {
-		t.Fatalf("startup kind = %q, want %q", got.Kind, startupcli.KindMCPServe)
-	}
-	if !got.MCP.RuntimeBacked || got.MCP.ProjectRoot != "project" {
-		t.Fatalf("startup MCP intent = %#v, want generated flag values", got.MCP)
-	}
-	if got.MCP.Stdin != stdin || got.MCP.Stdout != &stdout {
-		t.Fatalf("startup stdio = (%T, %T), want command streams", got.MCP.Stdin, got.MCP.Stdout)
-	}
-
-	workflow, _, err := root.Find([]string{"workflow"})
-	if err != nil {
-		t.Fatalf("find workflow: %v", err)
-	}
-	for _, name := range []string{"run", "start", "status", "result", "dispatches", "artifacts", "events"} {
-		if child, _, findErr := workflow.Find([]string{name}); findErr != nil || child.Name() != name {
-			t.Fatalf("out-of-scope workflow compatibility command %q missing after cutover: child=%v err=%v", name, child, findErr)
-		}
 	}
 }
